chore!: dropped support for legacy transxchange map processor
