+++ module/Api/src/Domain/Command/Bus/Ebsr/ProcessRequestMap.php
@@ -18,8 +18,6 @@ class ProcessRequestMap extends AbstractIdOnlyCommand
 
     protected $regNo;
 
-    protected $fromNewEbsr;
-
     /**
      * Gets the scale
      *
@@ -48,15 +46,5 @@ class ProcessRequestMap extends AbstractIdOnlyCommand
     public function getRegNo()
     {
         return $this->regNo;
-    }
-
-    /**
-     * Gets whether the request is a result of a new EBSR pack
-     *
-     * @return bool
-     */
-    public function getFromNewEbsr()
-    {
-        return $this->fromNewEbsr;
     }
 }
