+++ module/Api/src/Service/Ebsr/FileProcessor.php
@@ -56,7 +56,7 @@ class FileProcessor implements FileProcessorInterface
      * @throws \RuntimeException
      * @throws EbsrPackException
      */
-    public function fetchXmlFileNameFromDocumentStore($identifier, $isTransXchange = false)
+    public function fetchXmlFileNameFromDocumentStore($identifier)
     {
         $targetDir = $this->tmpDir . $this->subDirPath;
 
@@ -77,12 +77,6 @@ class FileProcessor implements FileProcessorInterface
             $this->decompressFilter->filter($filePath);
         } catch (LaminasFilterRuntimeException $e) {
             throw new EbsrPackException(self::DECOMPRESS_ERROR_PREFIX . $e->getMessage());
-        }
-
-        //transxchange runs through tomcat, therefore tomcat needs permissions on the files we've just created
-        if ($isTransXchange) {
-            $execCmd = 'setfacl -bR -m u:tomcat:rwx ' . $tmpDir;
-            exec(escapeshellcmd($execCmd));
         }
 
         $finder = new Finder();
