+++ test/module/Api/src/Domain/CommandHandler/Bus/Ebsr/ProcessPackTransactionTest.php
@@ -1407,7 +1407,6 @@ class ProcessPackTransactionTest extends ProcessPackTestCase
         $requestMapData = [
             'id' => $savedBusRegId,
             'scale' => 'auto',
-            'fromNewEbsr' => true
         ];
 
         $this->expectedSideEffect(RequestMapQueueCmd::class, $requestMapData, new Result());
